move tree rendition types to ui/tree.rs
