+++ src/task.rs
@@ -18,6 +18,7 @@ use itertools::Itertools;
 use crate::config::config_file::toml::TomlParser;
 use crate::config::Config;
 use crate::file;
+use crate::ui::tree::TreeItem;
 
 #[derive(Debug, Default, Clone, Eq, PartialEq)]
 pub struct Task {
@@ -287,76 +288,17 @@ impl Deps {
     // }
 }
 
-pub trait TreeItem: Clone {
-    type Child: TreeItem;
-
-    fn write_self(&self) -> Result<()>;
-
-    fn children(&self) -> Cow<[Self::Child]>;
-}
-
 impl TreeItem for (&Graph<Task, ()>, NodeIndex) {
     type Child = Self;
 
-    fn write_self(&self) -> Result<()> {
+    fn write_self(&self) {
         if let Some(w) = self.0.node_weight(self.1) {
             miseprint!("{}", w.name);
         }
-        Ok(())
     }
 
     fn children(&self) -> Cow<[Self::Child]> {
         let v: Vec<_> = self.0.neighbors(self.1).map(|i| (self.0, i)).collect();
         Cow::from(v)
-    }
-}
-
-pub struct TreeItemIndentChars {
-    /// Character for pointing down and right (`├`).
-    pub down_and_right: &'static str,
-    /// Character for pointing straight down (`|`).
-    pub down: &'static str,
-    /// Character for turning from down to right (`└`).
-    pub turn_right: &'static str,
-    /// Character for pointing right (`─`).
-    pub right: &'static str,
-    /// Empty character (` `).
-    pub empty: &'static str,
-}
-
-pub const TREE_ITEM_CHARS: TreeItemIndentChars = TreeItemIndentChars {
-    down_and_right: "├",
-    down: "│",
-    turn_right: "└",
-    right: "─",
-    empty: " ",
-};
-
-pub struct TreeItemIndent {
-    pub regular_prefix: String,
-    pub child_prefix: String,
-    pub last_regular_prefix: String,
-    pub last_child_prefix: String,
-}
-
-impl TreeItemIndent {
-    pub fn new(
-        indent_size: usize,
-        padding: usize,
-        characters: &TreeItemIndentChars,
-    ) -> TreeItemIndent {
-        let m = 1 + padding;
-        let n = if indent_size > m { indent_size - m } else { 0 };
-
-        let right_pad = characters.right.repeat(n);
-        let empty_pad = characters.empty.repeat(n);
-        let item_pad = characters.empty.repeat(padding);
-
-        TreeItemIndent {
-            regular_prefix: format!("{}{}{}", characters.down_and_right, right_pad, item_pad),
-            child_prefix: format!("{}{}{}", characters.down, empty_pad, item_pad),
-            last_regular_prefix: format!("{}{}{}", characters.turn_right, right_pad, item_pad),
-            last_child_prefix: format!("{}{}{}", characters.empty, empty_pad, item_pad),
-        }
     }
 }
